Changed interface of choose function to more accurately express the capabilities of the underlying algorithm

+++ include/moremath.h
@@ -1,7 +1,7 @@
 #ifndef MOREMATH_H
 #define MOREMATH_H
 
-double choose(double n, double k);
+int choose(int n, int k);
 double binomial_pdf(int x, double p, double n);
 double binomial_cdf(int x,double p, int n);
